Rename functions/methods and move InboxRuleHandlerTest
