+++ test/tests/Suite.ts
@@ -141,7 +141,7 @@ import "./gui/base/QrCodeScannerTest.js"
 import "./gui/base/WizardDialogNTest.js"
 import "./login/LoginViewModelTest.js"
 import "./login/PostLoginUtilsTest.js"
-import "./mail/InboxRuleHandlerTest.js"
+import "./mail/inboxrules/InboxRuleHandlerTest.js"
 import "./mail/ProcessInboxHandlerTest.js"
 import "./mail/KnowledgeBaseSearchFilterTest.js"
 import "./mail/MailModelTest.js"
